Added ADD_NEW_CLIENT and DELETE_CLIENT enum type of TypeOpertion, new class BankAck in Messaages package

+++ src/clients/Client.java
@@ -48,7 +48,7 @@ public class Client
      * Getter
      * @return id
      */
-    public int getId()
+    public long getId()
     {
         return id;
     }
